Refactor PersonalSchedulesDetailsPage and add Weekly View page
- Removed unused code and helper functions from PersonalSchedulesDetailsPage.
- Added a button to navigate to the new Weekly View page.
- Created PersonalSchedulesWeeklyViewPage to display the weekly schedule with event transformation logic.
